refactor: deduplicate method
