+++ src/main/kotlin/com/github/trc/clayium/api/metatileentity/multiblock/ClayBlastFurnaceMetaTileEntity.kt
@@ -10,8 +10,8 @@ import com.github.trc.clayium.api.capability.impl.NotifiableItemStackHandler
 import com.github.trc.clayium.api.metatileentity.MetaTileEntity
 import com.github.trc.clayium.api.metatileentity.WorkableMetaTileEntity
 import com.github.trc.clayium.api.metatileentity.multiblock.MultiblockLogic.StructureValidationResult
-import com.github.trc.clayium.api.util.CUtils.clayiumId
 import com.github.trc.clayium.api.util.ITier
+import com.github.trc.clayium.api.util.clayiumId
 import com.github.trc.clayium.common.recipe.registry.CRecipes
 import net.minecraft.client.renderer.block.model.ModelResourceLocation
 import net.minecraft.item.Item
