+++ src/main/kotlin/com/github/trc/clayium/client/renderer/ClayLaserReflectorRenderer.kt
@@ -1,7 +1,7 @@
 package com.github.trc.clayium.client.renderer
 
 import com.github.trc.clayium.api.capability.ClayiumTileCapabilities
-import com.github.trc.clayium.api.util.CUtils.clayiumId
+import com.github.trc.clayium.api.util.clayiumId
 import com.github.trc.clayium.common.blocks.BlockClayLaserReflector
 import com.github.trc.clayium.common.blocks.TileEntityClayLaserReflector
 import net.minecraft.client.renderer.GlStateManager
