+++ src/main/kotlin/com/github/trc/clayium/common/metatileentities/CaInjectorMetaTileEntity.kt
@@ -6,8 +6,8 @@ import com.github.trc.clayium.api.capability.impl.RecipeLogicCaInjector
 import com.github.trc.clayium.api.capability.impl.ResonanceManager
 import com.github.trc.clayium.api.metatileentity.MetaTileEntity
 import com.github.trc.clayium.api.metatileentity.WorkableMetaTileEntity
-import com.github.trc.clayium.api.util.CUtils.clayiumId
 import com.github.trc.clayium.api.util.ITier
+import com.github.trc.clayium.api.util.clayiumId
 import com.github.trc.clayium.common.recipe.registry.CRecipes
 import net.minecraft.util.ResourceLocation
