+++ src/main/kotlin/com/github/trc/clayium/common/metatileentities/CobblestoneGeneratorMetaTileEntity.kt
@@ -3,8 +3,8 @@ package com.github.trc.clayium.common.metatileentities
 import com.github.trc.clayium.api.CValues
 import com.github.trc.clayium.api.metatileentity.AbstractItemGeneratorMetaTileEntity
 import com.github.trc.clayium.api.metatileentity.MetaTileEntity
-import com.github.trc.clayium.api.util.CUtils.clayiumId
 import com.github.trc.clayium.api.util.ITier
+import com.github.trc.clayium.api.util.clayiumId
 import net.minecraft.block.material.Material
 import net.minecraft.client.renderer.block.model.ModelResourceLocation
 import net.minecraft.init.Blocks
